requirements.txt
pandas
numpy
scikit-learn
spacy
pdfminer.six
python-docx
streamlit
PyMuPDF
